pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = "MECHA"
version = "0.1.0"
description = "Parametric CAD + deterministic design helpers with optional AI assists for robotics/mechatronics."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Clark", email = "thungkao.kla@gmail.com" }]
dependencies = [
    "cadquery>=2.4.0,<2.6.0",
    "cadquery-ocp>=7.7.0,<7.8",
    "pyyaml>=6.0"
]

[project.scripts]
mmcad = "mmcad.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mm.ai]
provider = "cloud"         # "none" | "cloud"
model = "replace-me"       # e.g. "gpt-4.x" / "claude-3.x"
